Client and server - Sanitising inputs

+++ server/router/admin/shipment.js
@@ -1,11 +1,21 @@
 const express = require('express');
 const router = express.Router();
 const orderService = require("../../utils/orderService");
-const worker = require("../../utils/bgworkers/worker");
+const emailWorker = require("../../utils/bgworkers/worker");
+const { check, validationResult } = require('express-validator');
 
+router.post('/', [
+    check('userID').trim().escape(),
+    check('shipmentID').trim().escape(),
+    check('orderID').trim().escape(),
+    check('shipmentValue').isNumeric().trim().escape(),
+    check('shipmentWeight').isNumeric().trim().escape(),
+], (req, res, next) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(404).send({ error: 'Try again later' })
+    }
 
-router.post('/', (req, res, next) => {
-    //Sanitize the body
     orderService.addShipment(req.body)
         .then((result) => {
             if (result.error) {
@@ -13,15 +23,21 @@ router.post('/', (req, res, next) => {
             }
             let { shipment } = result;
             orderService.updateOrderStatus(shipment.orderID, "ORDER_SHIPMENT_ADDED");
-            worker.emailQueue.add({ status: "ORDER_SHIPMENT_ADDED", shipment });
+            emailWorker.emailQueue.add({ status: "ORDER_SHIPMENT_ADDED", shipment });
             return res.status(200).send({ shipment, message: result.message });
         })
         .catch(err => next(err));
 })
 
-router.post('/all', (req, res, next) => {
+router.post('/all', [
+    check('orderID').trim().escape()
+], (req, res, next) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(404).send({ error: 'Try again later' })
+    }
+
     let { orderID } = req.body;
-    //Sanitize
     orderService.getShipmentsFromOrder(orderID)
         .then(shipments =>
             shipments ? res.status(200).json(shipments) : res.status(404).send({ error: 'Try again later' })
@@ -29,9 +45,18 @@ router.post('/all', (req, res, next) => {
         .catch(err => next(err));
 })
 
-router.post('/update', (req, res, next) => {
+
+router.post('/update', [
+    check('shipmentID').trim().escape(),
+    check('shipmentValue').isNumeric().trim().escape(),
+    check('shipmentWeight').isNumeric().trim().escape(),
+], (req, res, next) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(404).send({ error: 'Try again later' })
+    }
+
     let { shipmentID, shipmentValue, shipmentWeight } = req.body;
-    //Sanitize
     orderService.updateShipment(shipmentID, shipmentValue, shipmentWeight)
         .then((shp) => {
             if (shp) {
@@ -41,9 +66,15 @@ router.post('/update', (req, res, next) => {
         .catch(err => next(err));
 })
 
-router.post('/delete', (req, res, next) => {
+router.post('/delete', [
+    check('shipmentID').trim().escape()
+], (req, res, next) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(404).send({ error: 'Try again later' })
+    }
+
     let { shipmentID } = req.body;
-    //Sanitize
     orderService.deleteShipment(shipmentID)
         .then(shipment =>
             shipment ? res.status(200).json({ message: "Shipment deleted" }) : res.status(404).send({ error: 'Try again later' })
